Switch to the new PatchHex function

+++ src/patch_tools.h
@@ -6,7 +6,7 @@
 
 //Return 0 on success, return 1 on failure.
 //4 HEX ie FF FF FF FF
-int PatchHex_32(void* baseaddress,uint64_t offset, uint32_t original_hex, uint32_t new_hex);
+int PatchHex_32(void* baseaddress,int offset, uint32_t original_hex, uint32_t new_hex);
 
 
 //1 HEX ie: FF
